+++ unstructured_inference/inference/elements.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-import os
 import re
 import unicodedata
 from copy import deepcopy
@@ -13,9 +12,7 @@ from scipy.sparse.csgraph import connected_components
 
 from unstructured_inference.config import inference_config
 from unstructured_inference.constants import Source
-from unstructured_inference.logger import logger
 from unstructured_inference.math import safe_division
-from unstructured_inference.models import tesseract
 
 
 @dataclass
@@ -208,24 +205,17 @@ class TextRegion(Rectangle):
         objects: Optional[Collection[TextRegion]],
         image: Optional[Image.Image] = None,
         extract_tables: bool = False,
-        ocr_strategy: str = "auto",
-        ocr_languages: str = "eng",
     ) -> str:
         """Extracts text contained in region."""
         if self.text is not None:
             # If block text is already populated, we'll assume it's correct
             text = self.text
         elif objects is not None:
-            text = aggregate_by_block(self, image, objects, ocr_strategy)
-        elif image is not None:
-            # We don't have anything to go on but the image itself, so we use OCR
-            text = ocr(self, image, languages=ocr_languages) if ocr_strategy != "never" else ""
+            text = aggregate_by_block(self, objects)
         else:
-            raise ValueError(
-                "Got arguments image and layout as None, at least one must be populated to use for "
-                "text extraction.",
-            )
-        return text
+            text = ""
+        cleaned_text = remove_control_characters(text)
+        return cleaned_text
 
 
 class EmbeddedTextRegion(TextRegion):
@@ -234,8 +224,6 @@ class EmbeddedTextRegion(TextRegion):
         objects: Optional[Collection[TextRegion]],
         image: Optional[Image.Image] = None,
         extract_tables: bool = False,
-        ocr_strategy: str = "auto",
-        ocr_languages: str = "eng",
     ) -> str:
         """Extracts text contained in region."""
         if self.text is None:
@@ -250,97 +238,22 @@ class ImageTextRegion(TextRegion):
         objects: Optional[Collection[TextRegion]],
         image: Optional[Image.Image] = None,
         extract_tables: bool = False,
-        ocr_strategy: str = "auto",
-        ocr_languages: str = "eng",
     ) -> str:
         """Extracts text contained in region."""
         if self.text is None:
-            if ocr_strategy == "never" or image is None:
-                return ""
-            else:
-                return ocr(self, image, languages=ocr_languages)
-        else:
-            return super().extract_text(objects, image, extract_tables, ocr_strategy)
-
-
-def ocr(text_block: TextRegion, image: Image.Image, languages: str = "eng") -> str:
-    """Runs a cropped text block image through and OCR agent."""
-    logger.debug("Running OCR on text block ...")
-    tesseract.load_agent(languages=languages)
-    padded_block = text_block.pad(12)
-    cropped_image = image.crop((padded_block.x1, padded_block.y1, padded_block.x2, padded_block.y2))
-    entrie_page_ocr = os.getenv("ENTIRE_PAGE_OCR", "tesseract").lower()
-    if entrie_page_ocr == "paddle":
-        from unstructured_inference.models import paddle_ocr
-
-        paddle_result = paddle_ocr.load_agent().ocr(np.array(cropped_image), cls=True)
-        recognized_text = ""
-        for idx in range(len(paddle_result)):
-            res = paddle_result[idx]
-            for line in res:
-                recognized_text += line[1][0]
-        return recognized_text
-    else:
-        agent = tesseract.ocr_agents.get(languages)
-        if agent is None:
-            raise RuntimeError("OCR agent is not loaded for {languages}.")
-
-        try:
-            return agent.detect(cropped_image)
-        except tesseract.TesseractError:
-            logger.warning("TesseractError: Skipping region", exc_info=True)
             return ""
-
-
-def needs_ocr(
-    region: TextRegion,
-    pdf_objects: Collection[TextRegion],
-    ocr_strategy: str,
-) -> bool:
-    """Logic to determine whether ocr is needed to extract text from given region."""
-    if ocr_strategy == "force":
-        return True
-    elif ocr_strategy == "auto":
-        image_objects = [obj for obj in pdf_objects if isinstance(obj, ImageTextRegion)]
-        word_objects = [obj for obj in pdf_objects if isinstance(obj, EmbeddedTextRegion)]
-        # If any image object overlaps with the region of interest, we have hope of getting some
-        # text from OCR. Otherwise, there's nothing there to find, no need to waste our time with
-        # OCR.
-        image_intersects = any(region.intersects(img_obj) for img_obj in image_objects)
-        if region.text is None:
-            # If the region has no text check if any images overlap with the region that might
-            # contain text.
-            if any(obj.is_in(region) and obj.text is not None for obj in word_objects):
-                # If there are word objects in the region, we defer to that rather than OCR
-                return False
-            else:
-                return image_intersects
         else:
-            # If the region has text, we should only have to OCR if too much of the text is
-            # uninterpretable.
-            return cid_ratio(region.text) > 0.5
-    else:
-        return False
+            return super().extract_text(objects, extract_tables)
 
 
 def aggregate_by_block(
     text_region: TextRegion,
-    image: Optional[Image.Image],
     pdf_objects: Collection[TextRegion],
-    ocr_strategy: str = "auto",
-    ocr_languages: str = "eng",
 ) -> str:
     """Extracts the text aggregated from the elements of the given layout that lie within the given
     block."""
-    if image is not None and needs_ocr(text_region, pdf_objects, ocr_strategy):
-        text = ocr(text_region, image, languages=ocr_languages)
-    else:
-        filtered_blocks = [obj for obj in pdf_objects if obj.is_in(text_region, error_margin=5)]
-        for little_block in filtered_blocks:
-            if image is not None and needs_ocr(little_block, pdf_objects, ocr_strategy):
-                little_block.text = ocr(little_block, image, languages=ocr_languages)
-        text = " ".join([x.text for x in filtered_blocks if x.text])
-    text = remove_control_characters(text)
+    filtered_blocks = [obj for obj in pdf_objects if obj.is_in(text_region, error_margin=5)]
+    text = " ".join([x.text for x in filtered_blocks if x.text])
     return text
 
 
@@ -363,6 +276,10 @@ def is_cid_present(text: str) -> bool:
 
 def remove_control_characters(text: str) -> str:
     """Removes control characters from text."""
+
+    # Replace newline character with a space
+    text = text.replace("\n", " ")
+    # Remove other control characters
     out_text = "".join(c for c in text if unicodedata.category(c)[0] != "C")
     return out_text